Moved the constants into the defaults structure and now can export the config before creating the repo
New constants structure includes most of the other previous constants. This structure can be exported before creating a repo by running config-init. The config file will also be created by default, if no config file is specified, it will default to the config constant.

Separated the main and pages branch into different command line options on their relevant commands - this was partially driven by the GitHub change, it will default to the current values GitHub uses by default right now but gives a bit more flexibility in what is used.

Fixed one of the tests that were failing due to the change. There is still one fail but not familiar with the libraries involved to resolve.